+++ app/data/testnet/knownAddressBranding.ts
@@ -1,4 +1,7 @@
 import type {KnownAddressBranding} from "../knownAddressBranding";
+import {aptosFrameworkAddressBranding} from "../aptosFrameworkAddressBranding";
 
 export const testnetKnownAddressBranding: Record<string, KnownAddressBranding> =
-  {};
+  {
+    ...aptosFrameworkAddressBranding,
+  };